dom/tests/mochitest/pointerlock/file_pointerlockerror.html

<!DOCTYPE HTML>
<html>
<!--https://bugzilla.mozilla.org/show_bug.cgi?id=633602-->
<head>
  <title>Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
  </script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="application/javascript" src="pointerlock_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>

  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * Make sure pointerlockerror event fires.
       */

      SimpleTest.waitForExplicitFinish();

      document.addEventListener("pointerlockerror", function (e) {
        ok(true, "pointerlockerror event should fire.");
        SimpleTest.finish();
      }, false);

      function start() {
        // element not in the DOM, not fullscreen, should fail to lock
        div = document.createElement("div");
        div.requestPointerLock();
      }
    </script>
  </pre>
</body>
</html>
